Move UriViewer into data table directory
