+++ src/components/data/data-utils.js
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { b, div, h, img } from 'react-hyperscript-helpers';
 import { absoluteSpinnerOverlay, ButtonPrimary, Clickable, DeleteConfirmationModal, Select, spinnerOverlay } from 'src/components/common';
 import Modal from 'src/components/Modal';
-import { isAzureUri, isGsUri } from 'src/components/UriViewer/uri-viewer-utils';
+import { isAzureUri, isGsUri } from 'src/data/data-table/uri-viewer/uri-viewer-utils';
 import ReferenceData from 'src/data/reference-data';
 import { Ajax } from 'src/libs/ajax';
 import { canUseWorkspaceProject } from 'src/libs/ajax/Billing';
