+++ src/data/data-table/entity-service/renderDataCell.js
@@ -6,12 +6,13 @@ import { parseGsUri } from 'src/components/data/data-utils';
 import { icon } from 'src/components/icons';
 import { TextCell } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
-import { isAzureUri, isDrsUri, isGsUri } from 'src/components/UriViewer/uri-viewer-utils';
-import { UriViewerLink } from 'src/components/UriViewer/UriViewerLink';
 import { getAttributeType } from 'src/data/data-table/entity-service/attribute-utils';
 import colors from 'src/libs/colors';
 import * as Utils from 'src/libs/utils';
 import { isAzureWorkspace, isGoogleWorkspace } from 'src/libs/workspace-utils';
+
+import { UriViewerLink } from '../shared/UriViewerLink';
+import { isAzureUri, isDrsUri, isGsUri } from '../uri-viewer/uri-viewer-utils';
 
 const isViewableUri = (datum, workspace) =>
   (isGoogleWorkspace(workspace) && isGsUri(datum)) || (isAzureWorkspace(workspace) && isAzureUri(datum)) || isDrsUri(datum);
