+++ src/data/data-table/uri-viewer/uri-viewer-utils.ts
